Refactor point representations
This especially effects the PointGrouper class. I had Point and PointPosition as two separate models which was dumb. This makes Point store its point in the array, so 0..<width or height instead of its point on screen. I also changed as many values as I could to be CGFloat to do less type conversion.