Optimizations to PointGrouper to skip scanned peaks
If there is a large region that is over the height threshold, it's very possible for it to contain multiple peaks. Since I'm using peaks as the start point for each group, scanning each peak in an area that was already scanned was wasteful. This marks points as "grouped" once they have been aded to a group. Peak points that have been grouped already will be skipped.