Fix a bug in the GroupWrapper
I keep assume Y is the top. This is becoming more and more of a problem.

The start point of the wrapper first looks in the up direction for a
point. However, the start point (which I've had a lot of trouble
finding correctly) was at the bottom, not the top. So instead of fixing
the troublesome start point, I just changed the initial direction to be
down, which seems to work!

Peaks.xcodeproj/project.xcworkspace/xcuserdata/edwardloveall.xcuserdatad
/UserInterfaceState.xcuserstate
Peaks.xcodeproj/xcuserdata/edwardloveall.xcuserdatad/xcschemes/Peaks.xcs
cheme